Homme: move util for toggling output out of Session
