+++ components/homme/src/share/cxx/Hommexx_Session.hpp
@@ -35,8 +35,6 @@ struct Session {
   }
 
   static bool m_screen_output_enabled;
-
-  static void toggle_screen_output (const bool enabled);
 };
 
 extern "C" {
