+++ components/homme/src/share/cxx/utilities/IOUtils.hpp
@@ -1,0 +1,17 @@
+/********************************************************************************
+ * HOMMEXX 1.0: Copyright of Sandia Corporation
+ * This software is released under the BSD license
+ * See the file 'COPYRIGHT' in the HOMMEXX/src/share/cxx directory
+ *******************************************************************************/
+
+#ifndef HOMMEXX_IO_UTILS_HPP
+#define HOMMEXX_IO_UTILS_HPP
+
+namespace Homme
+{
+
+void toggle_screen_output (const bool enabled);
+
+} // namespace Homme
+
+#endif // HOMMEXX_IO_UTILS_HPP
